requirements.txt
absl-py==2.1.0
asgiref==3.7.2
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.12.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==5.1.1
Django==5.0.1
django-extensions==3.2.3
dm-tree==0.1.8
docopt==0.6.2
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
Js2Py==0.74
keras==2.15.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
namex==0.0.7
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
openai-whisper==20231117
opt-einsum==3.3.0
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.1
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.6.0
torch==2.2.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
